Attempt to enclose unquoted strings
This is an attempt to enclose unquoted strings with quotes. Unquoted
"strings" that only consist of numbers (or decimal places) are ignored.
Trailing whitespace is trimmed from the enclosed string, which may or
may not be intended, but is generally impossible to detect when it
should be preserved. This also makes no attempt to account for strings
with missing quotes, and will probably break anything that consumes the
result of this call - but that would happen anyway.